Added tooltips, some reorganization
* Added two variants of Not Implemented tooltip for unfinished items
* "Released" unused but finished crafting items
* Added waterproofness to WireBlock (probably needs to use capabilities)
* Reorganized Blocks folder
