+++ src/main/java/com/arrl/radiocraft/common/blocks/antennas/SlimJimAntennaBlock.java
@@ -1,0 +1,21 @@
+package com.arrl.radiocraft.common.blocks.antennas;
+
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.TooltipFlag;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.List;
+
+public class SlimJimAntennaBlock extends AntennaCenterBlock {
+    public SlimJimAntennaBlock(Properties properties) {
+        super(properties);
+    }
+
+    @Override
+    public void appendHoverText(@NotNull ItemStack stack, Item.@NotNull TooltipContext context, @NotNull List<Component> tooltipComponents, @NotNull TooltipFlag tooltipFlag) {
+        super.appendHoverText(stack, context, tooltipComponents, tooltipFlag);
+        tooltipComponents.add(Component.translatable("tooltip.radiocraft.not_implemented"));
+    }
+}
