+++ src/main/java/com/arrl/radiocraft/common/blocks/antennas/YagiAntennaBlock.java
@@ -1,18 +1,25 @@
-package com.arrl.radiocraft.common.blocks;
+package com.arrl.radiocraft.common.blocks.antennas;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.state.BlockBehaviour;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.VoxelShape;
 import org.jetbrains.annotations.NotNull;
 
+import java.util.List;
+
 public class YagiAntennaBlock extends DoubleVHFAntennaBlock {
     public static final VoxelShape POST = Block.box(11.0D, 0.0D, 6.0D, 15.0D, 16.0D, 10.0D);
 
-    public YagiAntennaBlock(Properties properties) {
+    public YagiAntennaBlock(BlockBehaviour.Properties properties) {
         super(properties);
     }
 
@@ -24,5 +31,11 @@ public class YagiAntennaBlock extends DoubleVHFAntennaBlock {
     @Override
     protected @NotNull VoxelShape getCollisionShape(BlockState state, @NotNull BlockGetter level, @NotNull BlockPos pos, @NotNull CollisionContext context) {
         return rotateHorizontalPlaneDirection(Direction.WEST, state.getValue(FACING), POST);
+    }
+
+    @Override
+    public void appendHoverText(@NotNull ItemStack stack, Item.@NotNull TooltipContext context, @NotNull List<Component> tooltipComponents, @NotNull TooltipFlag tooltipFlag) {
+        super.appendHoverText(stack, context, tooltipComponents, tooltipFlag);
+        tooltipComponents.add(Component.translatable("tooltip.radiocraft.not_implemented"));
     }
 }
