+++ src/main/java/com/arrl/radiocraft/common/blocks/power/AbstractPowerNetworkBlock.java
@@ -1,5 +1,6 @@
-package com.arrl.radiocraft.common.blocks;
+package com.arrl.radiocraft.common.blocks.power;
 
+import com.arrl.radiocraft.common.blocks.AbstractNetworkBlock;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.InteractionResult;
