+++ src/main/java/com/arrl/radiocraft/common/init/RadiocraftTabs.java
@@ -22,10 +22,7 @@ public class RadiocraftTabs {
         // In the release version, we only enabled working items/blocks. So put all stuff not ready for release here.
         if (Radiocraft.IS_DEVELOPMENT_ENV) {
             output.accept(RadiocraftItems.MICROPHONE.get());
-            output.accept(RadiocraftItems.HF_CIRCUIT_BOARD.get());
             output.accept(RadiocraftItems.SMALL_BATTERY.get());
-            output.accept(RadiocraftItems.FERRITE_CORE.get());
-            output.accept(RadiocraftItems.COAXIAL_CORE.get());
             output.accept(RadiocraftItems.ANTENNA_ANALYZER.get());
             output.accept(RadiocraftItems.WATERPROOF_WIRE.get());
             output.accept(RadiocraftItems.SOLAR_PANEL.get());
@@ -58,6 +55,11 @@ public class RadiocraftTabs {
         }
 
         // Anything below here should be in the release creative menu
+
+        // General crafting-only ingredients (Not used yet, but they are "finished")
+        output.accept(RadiocraftItems.HF_CIRCUIT_BOARD.get());
+        output.accept(RadiocraftItems.FERRITE_CORE.get());
+        output.accept(RadiocraftItems.COAXIAL_CORE.get());
 
         // Crafting ingredients for VHF Handheld (the only actual working radio right now)
         output.accept(RadiocraftItems.RADIO_CRYSTAL.get());
