+++ src/main/java/com/arrl/radiocraft/compat/JadeCompat.java
@@ -1,13 +1,10 @@
 package com.arrl.radiocraft.compat;
 
-import com.arrl.radiocraft.common.blocks.AntennaCenterBlock;
-import com.arrl.radiocraft.common.init.RadiocraftBlocks;
+import com.arrl.radiocraft.common.blocks.antennas.AntennaCenterBlock;
 import snownee.jade.api.IWailaClientRegistration;
 import snownee.jade.api.IWailaCommonRegistration;
 import snownee.jade.api.IWailaPlugin;
 import snownee.jade.api.WailaPlugin;
-
-import java.util.ArrayList;
 
 @WailaPlugin
 public class JadeCompat implements IWailaPlugin {
